+++ packages/components/atoms/f-form-field/src/components/FormField.vue
@@ -24,6 +24,7 @@
                 :id="uniqueId"
                 :attributes="$attrs"
                 v-bind="$props"
+                :type="normalisedInputType"
                 v-on="listeners" />
 
             <form-dropdown